[fitting] move classes from fit to tools
